+++ src/site/markdown/release-notes.md
@@ -25,11 +25,21 @@
 
 ## Release Notes
 
+### 3.1.0
+<p class="smallinfo">Release date: 19 August 2020</p>
+
++ [HIPFORGE-363](https://issues.onehippo.com/browse/HIPFORGE-363)<br/> 
+  + added 'forwarded-for-header' setting on module level, see [usage page](usage.html).
+  + added 'cache-enabled' setting per configuration set, see [usage page](usage.html).
+  + more DEBUG logging in various locations, also `org.onehippo.forge.ipfilter.common.BaseIpFilter` will print out all 
+    request headers and cookies on TRACE level.  
+ 
+
 ### 3.0.1
 <p class="smallinfo">Release date: 30 January 2020</p>
 
 + [HIPFORGE-334](https://issues.onehippo.com/browse/HIPFORGE-334)<br/> 
-  IP-Filter: cache key should use password too.
+  the cache key should include password too.
 
 ### 3.0.0
 <p class="smallinfo">Release date: 7 February 2019</p>